Move administration, base & control unit use cases tests
